Allow drawing image in struct method by giving up genericity
The documentation encourages us to define method like draw<G: Graphics>(...),
but we have texture member variable, whose type depends on the generic
graphic type, so we need either generic Enemy type or concrete type for
draw method.  We chose the latter because we are writing concrete
application rather than a generic library.

This wasn't obvious because the type dependency was hidden in associated
type of Graphics::Texture.  I spent literally months to figure it out.