Add Assets struct to gather all assets
Now each Entity does not need to carry a reference to a texture,
which will make memory usage efficient.
Actual texture to be drawn is chosen by enum values, so it's part of
the type.

While there is certain benefit in putting a reference to a texture in
each Entity, the redundancy is less favorable as the program is getting
more complex.